# openapi.yaml
openapi: 3.1.0
info:
  title: Galaxy of Consequence RPG Backend API
  description: A comprehensive Star Wars RPG backend with NVIDIA Nemotron integration, persistent world simulation, reactive NPCs, and immersive Force mechanics following the Final Rule.
  version: 1.0.0
  contact:
    name: Galaxy of Consequence API
    email: api@galaxyofconsequence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.galaxyofconsequence.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /save_canvas:
    post:
      tags:
        - Canvas
      summary: Save player game state
      description: Save persistent game state to database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
              properties:
                user:
                  type: string
                  description: Player username
                campaign:
                  type: string
                  default: default
                canvas:
                  type: string
                  default: main
                data:
                  type: object
                  description: Game state data
                session_id:
                  type: string
                  description: Session identifier
      responses:
        '200':
          description: Canvas saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  canvas_id:
                    type: integer
                  timestamp:
                    type: string
  /get_canvas:
    get:
      tags:
        - Canvas
      summary: Get latest canvas for user
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: string
        - name: campaign
          in: query
          schema:
            type: string
            default: default
      responses:
        '200':
          description: Canvas retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasEntry'
  /get_canvas_by_id:
    get:
      tags:
        - Canvas
      summary: Get canvas by ID
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Canvas retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasEntry'
  /get_canvas_history:
    get:
      tags:
        - Canvas
      summary: Get canvas history
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: string
        - name: campaign
          in: query
          schema:
            type: string
            default: default
        - name: canvas
          in: query
          schema:
            type: string
            default: main
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Canvas history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/CanvasEntry'
                  total_entries:
                    type: integer
  /get_log:
    get:
      tags:
        - Canvas
      summary: Get filtered log entries
      parameters:
        - name: canvas
          in: query
          schema:
            type: string
        - name: user
          in: query
          schema:
            type: string
        - name: align
          in: query
          schema:
            type: string
            description: Force alignment filter
      responses:
        '200':
          description: Log entries retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  log_entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/CanvasEntry'
                  total_entries:
                    type: integer
                  filters_applied:
                    type: object

  # NVIDIA Nemotron Integration
...
  /faction_tick:
    post:
      tags:
        - Faction
      summary: Execute faction AI simulation
      description: Run autonomous faction AI turns for persistent world simulation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                force_tick:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Faction tick executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                  timestamp:
                    type: string
                  next_tick_in:
                    type: string
  /get_faction_state:
    get:
      tags:
        - Faction
      summary: Get faction state information
      parameters:
        - name: faction_name
          in: query
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Faction state retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  faction:
                    $ref: '#/components/schemas/FactionState'
                  factions:
                    type: array
                    items:
                      $ref: '#/components/schemas/FactionState'
  /update_faction_relationship:
    post:
      tags:
        - Faction
      summary: Update faction relationship with player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - faction_name
                - user
              properties:
                faction_name:
                  type: string
                user:
                  type: string
                relationship_change:
                  type: integer
                  default: 0
                awareness_change:
                  type: integer
                  default: 0
                action_description:
                  type: string
                session_id:
                  type: string
      responses:
        '200':
          description: Faction relationship updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  faction:
                    type: string
                  new_hostility:
                    type: integer
                  new_awareness:
                    type: integer
                  consequences:
                    type: array
                    items:
                      type: string

  # Quest System
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use 'Abracadabra' as the Bearer token
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /authenticate endpoint
  schemas:
...
    CanvasEntry:
      type: object
      properties:
        id:
          type: integer
        user:
          type: string
        campaign:
          type: string
        canvas:
          type: string
        data:
          type: object
        timestamp:
          type: string
        session_id:
          type: string
    FactionState:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        territory_control:
          type: integer
        resources:
          type: integer
        influence:
          type: integer
        awareness_level:
          type: integer
        hostility_level:
          type: integer
        active_operations:
          type: array
        strategic_goals:
          type: array
        last_action:
          type: string
...